Move the new command into interface/generate
