+++ interface/generate/templates/index_controller.jst
