+++ interface/generate/templates/package.json.jst
